src/pdf_assembler/requirements.txt
fpdf2>=2.7.0
Pillow>=10.0.0
boto3>=1.28.0
